+++ ownCloudSDK/HTTP/Request/OCHTTPRequest.m
@@ -380,6 +380,7 @@
 	BOOL readableContent = [contentType hasPrefix:@"text/"] ||
 			       [contentType isEqualToString:@"application/xml"] ||
 			       [contentType isEqualToString:@"application/json"] ||
+			       [contentType isEqualToString:@"application/jrd+json"] ||
 			       [contentType isEqualToString:@"application/x-www-form-urlencoded"];
 
 	NSString*(^FormatReadableData)(NSData *data) = ^(NSData *data) {
@@ -420,18 +421,39 @@
 {
 	NSMutableString *formattedHeaders = [NSMutableString new];
 
+	static dispatch_once_t onceToken;
+	static NSMutableArray<NSString *> *knownAuthorizationHeaders;
+
+	dispatch_once(&onceToken, ^{
+		knownAuthorizationHeaders = [NSMutableArray new];
+	});
+
 	[headers enumerateKeysAndObjectsUsingBlock:^(NSString * _Nonnull headerField, NSString * _Nonnull value, BOOL * _Nonnull stop) {
 		if ([headerField isEqualToString:OCHTTPHeaderFieldNameAuthorization])
 		{
 			NSArray<NSString *> *authorizationComponents = [value componentsSeparatedByString:@" "];
+			NSUInteger authHeaderIndex = 0;
+
+			@synchronized (knownAuthorizationHeaders)
+			{
+				// Do not store copies of header in the cache/memory - use its SHA-1 hash instead
+				NSString *hashedAuthorizationHeader = [[[value dataUsingEncoding:NSUTF8StringEncoding] sha1Hash] asHexStringWithSeparator:nil lowercase:YES];
+
+				if (![knownAuthorizationHeaders containsObject:hashedAuthorizationHeader])
+				{
+					[knownAuthorizationHeaders addObject:hashedAuthorizationHeader];
+				}
+
+				authHeaderIndex = [knownAuthorizationHeaders indexOfObject:hashedAuthorizationHeader];
+			}
 
 			if (authorizationComponents.count > 1)
 			{
-				value = [authorizationComponents[0] stringByAppendingString:@" [redacted]"];
+				value = [authorizationComponents[0] stringByAppendingFormat:@" [redacted:%lu]", authHeaderIndex];
 			}
 			else
 			{
-				value = @"[redacted]";
+				value = [NSString stringWithFormat:@"[redacted:%lu]", authHeaderIndex];
 			}
 		}
 
@@ -548,6 +570,8 @@
 		self.bodyData 		= [decoder decodeObjectOfClass:[NSData class] forKey:@"bodyData"];
 		self.bodyURL 		= [decoder decodeObjectOfClass:[NSURL class] forKey:@"bodyURL"];
 
+		self.authenticationDataID = [decoder decodeObjectOfClass:NSString.class forKey:@"authenticationDataID"];
+
 		self.redirectPolicy	= [decoder decodeIntegerForKey:@"redirectPolicy"];
 		self.maximumRedirectionDepth = [decoder decodeIntegerForKey:@"maximumRedirectionDepth"];
 		self.redirectionHistory = [decoder decodeObjectOfClasses:[[NSSet alloc] initWithObjects:NSMutableDictionary.class, NSURL.class, nil] forKey:@"redirectionHistory"];
@@ -600,6 +624,8 @@
 	[coder encodeObject:_headerFields 	forKey:@"headerFields"];
 	[coder encodeObject:_bodyData 		forKey:@"bodyData"];
 	[coder encodeObject:_bodyURL 		forKey:@"bodyURL"];
+
+	[coder encodeObject:_authenticationDataID forKey:@"authenticationDataID"];
 
 	[coder encodeInteger:_redirectPolicy 		forKey:@"redirectPolicy"];
 	[coder encodeInteger:_maximumRedirectionDepth 	forKey:@"maximumRedirectionDepth"];